UnityARPlayer.jar file to remove curl shared lib reference that is no longer needed
