+++ .gitignore
@@ -15,6 +15,9 @@ src/Unity/[Bb]uild/
 src/Unity/UnityGenerated/
 src/Unity/Assets/AssetStoreTools*
 
+src/Android_Unity_Player_Source/**/bin
+src/Android_Unity_Player_Source/EclipseProj/**/.settings
+
 # Autogenerated VS/MD solution and project files
 *.csproj
 *.unityproj
