+++ src/Android_Unity_Player_Source/AndroidStudioProj/readme.md
@@ -1,0 +1,1 @@
+Corresponding Android Studio project coming soon.
