+++ src/Android_Unity_Player_Source/EclipseProj/res/values/strings.xml
