+++ src/Android_Unity_ResLib_AS_Proj/.gitignore
@@ -1,0 +1,6 @@
+*.iml
+.gradle
+/local.properties
+/.idea/workspace.xml
+/.idea/libraries
+/captures
