+++ src/Android_Unity_ResLib_AS_Proj/.idea/.name
@@ -1,0 +1,1 @@
+Android_Unity_ResLib_AS_Proj
