+++ src/Android_Unity_ResLib_AS_Proj/.idea/modules.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectModuleManager">
+    <modules>
+      <module fileurl="file://$PROJECT_DIR$/Android_Unity_ResLib_AS_Proj.iml" filepath="$PROJECT_DIR$/Android_Unity_ResLib_AS_Proj.iml" />
+      <module fileurl="file://$PROJECT_DIR$/ResLib/ResLib.iml" filepath="$PROJECT_DIR$/ResLib/ResLib.iml" />
+    </modules>
+  </component>
+</project>
