+++ src/Android_Unity_ResLib_AS_Proj/ResLib/.gitignore
@@ -1,0 +1,1 @@
+/build
