+++ src/Android_Unity_ResLib_AS_Proj/ResLib/src/main/AndroidManifest.xml
@@ -1,0 +1,11 @@
+<manifest package="com.daqri.Unity.ResLib"
+          xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <application
+        android:allowBackup="true"
+        android:label="@string/app_name"
+        android:supportsRtl="true"
+        >
+    </application>
+
+</manifest>
