+++ src/Android_Unity_ResLib_AS_Proj/ResLib/src/main/res/values/strings.xml
@@ -1,0 +1,23 @@
+<resources>
+    <string name="app_name">ARTK_for_Unity_on_Android_Lib</string>
+        <string name="menuItem_settings">Settings</string>
+
+        <string name="prefCategory_title_camera">Camera settings</string>
+        <string name="pref_title_cameraResolution">Camera resolution</string>
+        <string-array name="pref_entries_cameraResolution">
+            <item>320x240   (4:3)</item>
+        </string-array>
+        <string-array name="pref_entryValues_cameraResolution">
+            <item>320x240</item>
+        </string-array>
+        <string name="pref_defaultValue_cameraResolution">320x240</string>
+
+        <string name="pref_title_cameraIndex">Change camera</string>
+        <string-array name="pref_entries_cameraIndex">
+            <item>Rear camera</item>
+        </string-array>
+        <string-array name="pref_entryValues_cameraIndex">
+            <item>0</item>
+        </string-array>
+
+    </resources>
