+++ src/Android_Unity_ResLib_AS_Proj/settings.gradle
@@ -1,0 +1,1 @@
+include ':ResLib'
