+++ src/Unity/Assets/Plugins/Android/ResLib-release.aar.meta
@@ -1,0 +1,16 @@
+fileFormatVersion: 2
+guid: 9dda4bc7975694c84baeb0f998cd40d3
+timeCreated: 1449517822
+licenseType: Pro
+PluginImporter:
+  serializedVersion: 1
+  iconMap: {}
+  executionOrder: {}
+  isPreloaded: 0
+  platformData:
+    Any:
+      enabled: 1
+      settings: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
