--- src/Unity/Assets/Plugins/Android/res.meta
@@ -1,5 +1,0 @@
-fileFormatVersion: 2
-guid: 6b384a31f3d914e898b606d71f83bd47
-folderAsset: yes
-DefaultImporter:
-  userData:
