--- src/Unity/Assets/Plugins/Android/res/values.meta
@@ -1,5 +1,0 @@
-fileFormatVersion: 2
-guid: 5b30bb034aae04957a6f440c53a36f87
-folderAsset: yes
-DefaultImporter:
-  userData:
