--- src/Unity/Assets/Plugins/Android/res/values/strings.xml.meta
@@ -1,4 +1,0 @@
-fileFormatVersion: 2
-guid: e9283d2b45ac2465782e96017b3657de
-DefaultImporter:
-  userData:
