--- src/Unity/Assets/Plugins/Android/res/xml.meta
@@ -1,5 +1,0 @@
-fileFormatVersion: 2
-guid: 60d55060fc1c341fe9345aca43e16758
-folderAsset: yes
-DefaultImporter:
-  userData:
