--- src/Unity/Assets/Plugins/Android/res/xml/preferences.xml.meta
@@ -1,4 +1,0 @@
-fileFormatVersion: 2
-guid: a66d310f755f04db1aa413a298d21be8
-DefaultImporter:
-  userData:
